feat: Add ViewModel implementation to Job Boards View
