+++ CIRN-2025/Job Board/JobBoardViewModel.swift
@@ -1,0 +1,14 @@
+//
+
+import Foundation
+
+class JobBoardViewModel {
+	
+	let jobs = [
+		"Chicago Steel Works",
+		"Urban Metals Inc.",
+		"West Side Gears",
+		"Lakeview Tool & Die",
+		"Midwest Fabrication Co."
+	]
+}
